+++ R/generic_functions.R
@@ -497,7 +497,7 @@ compute_stationary_variance <- function(variance, selection.strength){
     kro_sum_A_inv <- solve(kro_sum_A)
     root_var_vec <- kro_sum_A_inv %*% variance_vec
     gamma <- matrix(root_var_vec, dim(variance))
-    if (!isSymmetric(gamma)) stop("Error in computation of stationary variance: matrix computed was not symmetric.")
+    if (!isSymmetric(gamma, tol = (.Machine$double.eps)^(0.7))) stop("Error in computation of stationary variance: matrix computed was not symmetric.")
     gamma <- forceSymmetric(gamma)
     return(as(gamma, "dpoMatrix"))
   }